Seperates atoms out
